move processing status to common component
